feat: add Geist font and update image handling in setup API
- Added 'Geist' font to the Nuxt configuration.
- Updated package dependencies for @iconify-json/token, @nuxt/icon, and @types/bun.
- Modified setup API to accept multiple images with validation for size and format.
- Changed image handling to support an array of images instead of a single image.
- Improved error handling and logging for image uploads.
