+++ package.json
@@ -20,11 +20,11 @@
         "@iconify-json/mingcute": "^1.2.3",
         "@iconify-json/simple-icons": "^1.2.31",
         "@iconify-json/svg-spinners": "^1.2.2",
-        "@iconify-json/token": "^1.2.15",
+        "@iconify-json/token": "^1.2.16",
         "@iconify-json/twemoji": "^1.2.2",
         "@nuxt/eslint": "^1.3.0",
         "@nuxt/fonts": "^0.11.1",
-        "@nuxt/icon": "1.11.0",
+        "@nuxt/icon": "1.12.0",
         "@nuxt/image": "^1.10.0",
         "@nuxt/scripts": "^0.11.5",
         "@nuxtjs/color-mode": "3.5.2",
@@ -33,7 +33,7 @@
         "@nuxtjs/supabase": "1.5.0",
         "@nuxtjs/turnstile": "^1.0.0",
         "@tailwindcss/typography": "^0.5.16",
-        "@types/bun": "^1.2.8",
+        "@types/bun": "^1.2.9",
         "@types/canvas-confetti": "^1.9.0",
         "@vue/eslint-config-prettier": "^10.2.0",
         "@vueuse/nuxt": "^13.1.0",
